pyproject.toml
[tool.poetry]
name = "ad-progression-pipeline"
version = "0.1.0"
description = ""
authors = ["Minh <tminh.us@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
imbalanced-learn = "^0.12.3"
tensorflow = "2.17.0"
matplotlib = "^3.9.2"
numpy = "1.26.0"
pandas = "^2.2.2"
scikit-learn = "==1.0.2"
optuna = "^3.6.1"
pre-commit = "^3.8.0"
ruff = "^0.6.3"
mypy = "^1.11.2"
tensorflow-io-gcs-filesystem = "0.31.0"
prefect = "^2.20.4"
pyyaml = "^6.0.2"
scipy = "^1.14.1"
tbtrim = "^0.3.1"
tensorflow-cpu = "^2.17.0"
tensorflow-intel = "^2.12.0"
xgboost = "^2.1.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[tool.mypy]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 150
# Basic linting (PEP 8 compliance, etc.)
lint.select = [
    "ALL"
]
lint.ignore = [
    'D', # docstring shit
    'PD015',
    'ANN401',
    'N806',
    'G004', # logging statement f string
    'DTZ', # datetime bs
    'PTH',
    'FA',
    'FIX',
    'TD',
    'F4',
    'PD901'
]
